Cleanup Dependencies (#34)
* Remove `FillArrays`, `LogDensityProblemsAD`, `Requires`, `Accessors`

`LogDensityProblemsAD` and `Requires` are not obsolete. `Accessors` is used only in one place for mutating an array, so it can be removed. 

The removal `FillArrays` will increase memory consumption for univariate slice samplers since we now maintain `d` copies of the univariate slice sampler object. However, since slice samplers are not expected to work for large `d`, this should be fine.